+++ bugsnag-android-core/src/test/java/com/bugsnag/android/EventRedactionTest.kt
@@ -14,7 +14,7 @@ internal class EventRedactionTest {
         val event = Event(
             null,
             generateImmutableConfig(),
-            HandledState.newInstance(HandledState.REASON_HANDLED_EXCEPTION),
+            SeverityReason.newInstance(SeverityReason.REASON_HANDLED_EXCEPTION),
             NoopLogger
         )
         event.app = generateAppWithState()