+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## TBD
+
+* Make `event.unhandled` overridable for JVM errors
+  [#1025](https://github.com/bugsnag/bugsnag-android/pull/1025)
+
 ## 5.3.0 (2020-12-02)
 
 * Add integrity header to verify Error and Session API payloads have not changed